Animation handling in Character superclass
I'm thinking we should instantiate every possible character sprite in the constructor, and store it all in a list of lists, called animation_states. This way we won't need to use load_sprites every time the character walks and stops, we only switch which set of images it uses. So we change animation_current whenever we change the characters state and use that as input in the animate method, which is called every iteration of update.